// fso9/flight-diaries/flight-diary-frontend/src/types.ts
export interface DiaryEntry {
  id: number;
  date: string;
  weather: string;
  visibility: string;
  comment: string;
}

export type NewDiaryEntry = Omit<DiaryEntry, 'id'>;

export type NonSensitiveDiaryEntry = Omit<DiaryEntry, 'comment'>;

export interface ValidationError {
  message: string;
  errors: Record<string, string[]>
}